+++ frontend/src/components/AppSelectTags.vue
@@ -170,10 +170,10 @@ export type Option = {
 <script setup lang="ts">
 import { computed, nextTick, ref, watch } from "vue";
 import { isEqual, uniqBy, uniqueId } from "lodash";
+import { useFloating } from "@/composables/use-floating";
+import { useQuery } from "@/composables/use-query";
 import { wrap } from "@/util/math";
 import { copyToClipboard } from "@/util/string";
-import { useFloating } from "@/util/use-floating";
-import { useQuery } from "@/util/use-query";
 import AppInput from "./AppInput.vue";
 import { snackbar } from "./TheSnackbar.vue";
