+++ frontend/src/pages/PageHome.vue
@@ -192,8 +192,8 @@ import AppHighlight from "@/components/AppHighlight.vue";
 import AppPost from "@/components/AppPost.vue";
 import AppTabs from "@/components/AppTabs.vue";
 import AppTile from "@/components/AppTile.vue";
+import { useQuery } from "@/composables/use-query";
 import { formatNumber } from "@/util/string";
-import { useQuery } from "@/util/use-query";
 import tabs from "./explore/tabs.json";
 import TabSearch from "./explore/TabSearch.vue";
 import metadata from "./metadata.json";
