+++ frontend/src/pages/explore/TabTextAnnotator.vue
@@ -85,9 +85,9 @@ import { annotateText } from "@/api/text-annotator";
 import AppNodeBadge from "@/components/AppNodeBadge.vue";
 import AppTextbox from "@/components/AppTextbox.vue";
 import AppUpload from "@/components/AppUpload.vue";
+import { useQuery } from "@/composables/use-query";
 import { appendToBody } from "@/global/tooltip";
 import { downloadJson } from "@/util/download";
-import { useQuery } from "@/util/use-query";
 import example from "./text-annotator.json";
 
 /** text content */
