+++ frontend/src/pages/node/PageNode.vue
@@ -40,10 +40,10 @@ import { onMounted, watch } from "vue";
 import { useRoute } from "vue-router";
 import { getNode } from "@/api/node";
 import TheTableOfContents from "@/components/TheTableOfContents.vue";
+import { useQuery } from "@/composables/use-query";
 import { addEntry } from "@/global/history";
 import { appDescription, appTitle } from "@/global/meta";
 import SectionBreadcrumbs from "@/pages/node/SectionBreadcrumbs.vue";
-import { useQuery } from "@/util/use-query";
 import SectionAssociations from "./SectionAssociations.vue";
 import SectionExtra from "./SectionExtra.vue";
 import SectionHierarchy from "./SectionHierarchy.vue";
